backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.27.0
reportlab>=4.2.0
python-multipart>=0.0.9
pydantic>=2.10.0
bcrypt>=4.2.0
